chore: Add clearlog script to package.json
This commit adds a new script, "clearlog", to the package.json file. The "clearlog" script runs the "node ./util/clearlog.js" command. This script is useful for clearing log files in the application.

Note: The commit message has been generated based on the provided code changes and recent repository commits.